Add improved authorization to user management. Remove API-key authentication.

+++ src/main/java/dev/victorhleme/mobiauto/services/impl/UsuarioServiceImpl.java
@@ -33,10 +33,13 @@ public class UsuarioServiceImpl implements UsuarioService {
     private final RevendaService revendaService;
     private final EmailService emailService;
     private final PasswordEncoder passwordEncoder;
+    private final PermissionService permissionService;
 
     @Override
     public Usuario save(UsuarioCreationDto usuarioDto) {
+        permissionService.checkAuthority("PERM_USUARIO_CREATE_" + usuarioDto.getRevendaId());
         Usuario newUsuario = usuarioMapper.from(usuarioDto);
+
         newUsuario.setSenha(passwordEncoder.encode(usuarioDto.getSenha()));
 
         if (usuarioDto.getRevendaId() != null)
@@ -49,26 +52,37 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Override
     public Page<Usuario> getAll(UsuarioFilter filter) {
+        permissionService.checkAuthority("PERM_USUARIO_READ_" + filter.getRevendaId());
         Pageable pageable = PageRequest.of(filter.getPage(), filter.getSize());
         return usuarioRepository.findAll(usuarioSpecifications.getSpecification(filter), pageable);
     }
 
     @Override
     public Usuario findById(Long id) {
-        return usuarioRepository.findById(id)
-            .orElseThrow(() -> new NotFoundException(Usuario.class, id));
+        permissionService.checkAuthority("PERM_USUARIO_READ_" + getRevendaIdFromUser(id));
+        return findUsuarioOrThrowException(id);
     }
 
     @Override
     public Usuario update(UsuarioDto usuarioDto) {
-        Usuario existing = usuarioRepository.findById(usuarioDto.getId())
-            .orElseThrow(() -> new NotFoundException(Usuario.class, usuarioDto.getId()));
+        Usuario existing = findUsuarioOrThrowException(usuarioDto.getId());
+        permissionService.checkAuthority("PERM_USUARIO_MANAGE_" + existing.getRevenda().getId());
         BeanUtils.copyProperties(usuarioDto, existing);
         return usuarioRepository.save(existing);
     }
 
     @Override
     public void delete(Long id) {
+        permissionService.checkAuthority("PERM_USUARIO_MANAGE_" + getRevendaIdFromUser(id));
         usuarioRepository.deleteById(id);
+    }
+
+    private Usuario findUsuarioOrThrowException(Long id) {
+        return usuarioRepository.findById(id)
+            .orElseThrow(() -> new NotFoundException(Usuario.class, id));
+    }
+
+    private Long getRevendaIdFromUser(Long id) {
+        return findUsuarioOrThrowException(id).getRevenda().getId();
     }
 }
